WIP: Making the Cart queries run - broken test
The insert data is not getting populated in the database, although it is reading from the flat file and executing them. But, when I am trying to insert the data in POST Cart then, it was getting null data for the user and item. Here, the test cases are failing.